requirements.txt
python-telegram-bot==20.0a2
asyncpg==0.25.0
git+https://github.com/python-telegram-bot/ptbcontrib.git@main